HBSD: Disable RETPOLINE for ports-mgmt/pkg
It appears that FreeBSD's usage of ifuncs in libc causes issues with
pkg-static when it's compiled with RETPOLINE. I'm unsure of the
underlying cause, so more investigation is needed.

One possible cause is our use of the retpolineplt linker flag. Due to
libc being statically compiled in, it might be possible that using
retpolineplt causes issues with libc's PLT/GOT (statically-linked
binaries still have PLTs/GOTs for statically-linked dynamic libraries).

To reiterate, more investigation is needed.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Sponsored-by:	SoldierX
github-issue:	#164