HBSD: remove LibreSSL patch from net-mgmt/zabbix34-server
LibreSSL ports are marked broken and now the patch is broken
too, breaking the port build for everyone.

In a similar episode in OPNsense we learned that crypto patches
should best be pulled in via EXTRA_PATCHES only when they are
required, because otherwise they affect build stability and
sometimes even runtime stability in a negative way.

Signed-off-by: Franco Fichtner <franco.fichtner@hardenedbsd.org>