HBSD: Teach the `check-plist` target about HardenedBSD sanity checks
In upcoming an upcoming HardenedBSD Ports Contributor's Guide, I will
document the requirement of running `make check-plist` to ensure the
sanity of the unique HardenedBSD features.

Two hardening features are currently being sanity checked:

1. SafeStack
2. CFI

If a port as SafeStack and/or CFI enabled, it must not install shared
objects. Thus, if a port has SafeStack and/or CFI enabled, error out if
the port does indeed install shared objects.

Note that this commit needs to be followed up with a cleanup of those
ports in violation of the above rules (shells/zsh is one of them).

Look for the HardenedBSD Ports Contributor's Guide soon. It will be
posted on the HardenedBSD Ports wiki.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>